+++ mips/docs/sphinx/source/matlab-source/mips/have_feature_mips.m
@@ -1,0 +1,1 @@
+../../../../../lib/have_feature_mips.m