+++ mips/lib/mipsver.m
@@ -16,7 +16,7 @@ function rv = mipsver(varargin)
 % See also mpver.
 
 %   MIPS
-%   Copyright (c) 2010-2024, Power Systems Engineering Research Center (PSERC)
+%   Copyright (c) 2010-2025, Power Systems Engineering Research Center (PSERC)
 %   by Ray Zimmerman, PSERC Cornell
 %
 %   This file is part of MIPS.
@@ -24,9 +24,9 @@ function rv = mipsver(varargin)
 %   See https://github.com/MATPOWER/mips for more info.
 
 v = struct( 'Name',     'MIPS', ... 
-            'Version',  '1.5.1', ...
+            'Version',  '1.5.2', ...
             'Release',  '', ...
-            'Date',     '10-May-2024' );
+            'Date',     '05-Jul-2025' );
 if nargout > 0
     if nargin > 0
         rv = v;